refactor: changes structure of the app
