+++ src/app/(app)/@library/page.tsx
@@ -1,4 +1,4 @@
-import AlbumLibrary from '@/app/(app)/@albums/album-library';
+import AlbumLibrary from '@/app/(app)/@library/album-library';
 import Button from '@/components/ui/button';
 import Link from 'next/link';
 
@@ -9,7 +9,7 @@ const Albums = ({ searchParams }: { searchParams: { library: string } }) => {
 
   if (!library) {
     return (
-      <aside className='h-full'>
+      <aside className='h-full mobile:h-fit'>
         <Button
           variant='secondary'
           className='flex h-full flex-col items-center gap-2 mobile:h-fit mobile:flex-row mobile:items-center mobile:justify-center'
