+++ src/app/(app)/error.tsx
@@ -14,9 +14,5 @@ export default function Error({ error }: Props) {
     }
   }, [error]);
 
-  return (
-    <>
-      <h2>Something went wrong!</h2>
-    </>
-  );
+  return;
 }
