+++ src/app/layout.tsx
@@ -7,6 +7,7 @@ import { Toaster } from 'sonner';
 import '@/globals.css';
 import Navbar from '@/components/navbar';
 import Footer from '@/components/footer';
+import Providers from '@/app/providers';
 
 interface Props {
   children: ReactNode;
@@ -20,27 +21,29 @@ const metadata: Metadata = {
 
 const RootLayout = async ({ children }: Props) => {
   return (
-    <html lang='en' className='bg-black text-white'>
-      <body
-        className={`${GeistSans.className} flex h-dvh flex-col gap-4 overscroll-none p-4`}
-      >
-        <Navbar />
-        {children}
-        <Footer />
-        <Toaster
-          theme='dark'
-          position='bottom-center'
-          richColors
-          className='toaster group'
-          toastOptions={{
-            classNames: {
-              toast:
-                'group toast group-[.toaster]:bg-black group-[.toaster]:border-gray-100',
-            },
-          }}
-        />
-      </body>
-    </html>
+    <Providers>
+      <html lang='en' className='bg-black text-white'>
+        <body
+          className={`${GeistSans.className} flex h-dvh flex-col gap-4 overscroll-none p-4`}
+        >
+          <Navbar />
+          {children}
+          <Footer />
+          <Toaster
+            theme='dark'
+            position='bottom-center'
+            richColors
+            className='toaster group'
+            toastOptions={{
+              classNames: {
+                toast:
+                  'group toast group-[.toaster]:bg-black group-[.toaster]:border-gray-100',
+              },
+            }}
+          />
+        </body>
+      </html>
+    </Providers>
   );
 };
